fix(types): index ray_type_sizes by uint8_t to avoid OOB on atom tags
ray_elem_size(t) expanded to ray_type_sizes[(t)] with t an int8_t type tag.
For a negative (atom) tag — e.g. -RAY_BOOL == -1 — that is ray_type_sizes[-1],
an out-of-bounds read (UBSan: "index -1 out of bounds for type
'const uint8_t[256]'"), reached via ray_sym_elem_size() in the SYM-column read
path (observed in rfl/query/legacy_strl_predicates).

Cast the index to uint8_t, matching the indexing already used in vec.c: a
negative tag wraps into the upper, zero-filled half of the 256-entry table, so
atom tags read back a well-defined 0 (they carry no element size here) with no
OOB. The two direct ray_type_sizes[type] reads (eval.c gather, serde.c vector
decode) only ever see positive vector tags, so they are left as-is.

Extends test_types.c: ray_elem_size of negative / atom tags returns 0.

+++ src/core/types.h
@@ -39,8 +39,12 @@
 /* Type sizes lookup table (defined in types.c) */
 extern const uint8_t ray_type_sizes[256];
 
-/* Element size for a given type tag */
-#define ray_elem_size(t)  (ray_type_sizes[(t)])
+/* Element size for a given type tag.  Cast to uint8_t so a negative (atom)
+ * type tag indexes the upper, zero-filled half of the 256-entry table rather
+ * than reading out of bounds — ray_type_sizes[(int8_t)-1] is an OOB read (UB).
+ * Non-vector/atom tags legitimately have no element size and read back 0, and
+ * this matches the (uint8_t) indexing already used in vec.c. */
+#define ray_elem_size(t)  (ray_type_sizes[(uint8_t)(t)])
 
 static inline int64_t ray_cast_f64_to_i64_null(double v) {
     if (v != v) return NULL_I64;
